+++ tests/_internal/clients/test_alert_rules.py
@@ -3,7 +3,7 @@ import json
 import pytest
 
 from py42._internal.clients.detection_list_user import DetectionListUserClient
-from py42.clients.alertrules import AlertRulesClient
+from py42._internal.clients.alertrules import AlertRulesClient
 
 MOCK_DETECTION_LIST_GET_RESPONSE = """
 {"type$": "USER_V2", "tenantId": "1d71796f-af5b-4231-9d8e-df6434da4663",